### 2021-08-05 #3564 @Kevin-Mattheus-Moerman
@openjournals/dev can you advise how to fix this compilation error :point_up_2:
