### 2021-08-05 #3564 @danielskatz
My guess, from the error message, is that there's a problem with an svg image - since I see 2 such errors, and I see two svg images in https://github.com/GRChombo/GRChombo/tree/feature/joss_paper/figures, my thought is that JOSS doesn't handle svg images and that these images could be converted to pngs by the authors.
