### 2021-08-05 #3564 @mirenradia
This is rather strange since the [GitHub action run](https://github.com/GRChombo/GRChombo/actions/runs/1093660595) to build a preview of the paper passed successfully as well as the docker image. Is it possible to make these consistent with the actual PDF generator?

For now, I will convert the SVG images to PNGs.
