### 2021-08-05 3c183f1
Convert SVG figures to PNGs
