### 2021-08-05 #3564 @mirenradia
> For now, I will convert the SVG images to PNGs.

Done.
