### 2021-08-05 cb7863a
Remove irreducible mass from 2D runs and fix restart bug for merger AHs
Bug: when restarting from an early checkpoint file, SmallDataIO deletes extra lines in the stats file via 'remove_duplicate_time_data'. For mergers that do not necessarily start right after the restart, calling this does not work and the lines were never deleted. Solution: force SmallDataIO to think 'restart_time' is the current time when mergers solve for the first time, and hence allowing to remove duplicate lines.
