+++ Source/ApparentHorizonFinder/ApparentHorizon.impl.hpp
@@ -54,10 +54,13 @@ ApparentHorizon<SurfaceGeometry, AHFunction>::ApparentHorizon(
 #endif
       }),
 
-      m_area(NAN), m_mass(NAN),
+      m_area(NAN),
+#if GR_SPACEDIM == 3
+      m_mass(NAN), m_linear_momentum_P_norm(NAN),
 #if CH_SPACEDIM == 3
-      m_spin(NAN), m_irreducible_mass(NAN), m_spin_z_alt(NAN),
-      m_dimensionless_spin_vector({NAN}),
+      m_irreducible_mass(NAN), m_spin(NAN), m_spin_z_alt(NAN),
+      m_dimensionless_spin_vector({NAN}), m_linear_momentum_P({NAN}),
+#endif
 #endif
 
       origin_already_updated(false),
@@ -313,7 +316,7 @@ void ApparentHorizon<SurfaceGeometry, AHFunction>::solve(double a_dt,
     {
         if (m_params.verbose > AHParams::SOME)
         {
-            pout() << "In [ApparentHorizon::solve::post-solving]" << endl;
+            pout() << "In [ApparentHorizon::solve::post-solving]" << std::endl;
         }
 
         CH_TIME("ApparentHorizon::solve::post-solving");
@@ -347,15 +350,27 @@ void ApparentHorizon<SurfaceGeometry, AHFunction>::solve(double a_dt,
 
         origin_already_updated = false;
 
-        m_area = calculate_area();
 #if GR_SPACEDIM == 3 // GR_SPACEDIM, not CH_SPACEDIM !!!
 #if CH_SPACEDIM == 3
-        Tensor<1, double> J = calculate_angular_momentum_J();
+        Tensor<1, double> J;
+        calculate_ah_quantities(m_area, m_linear_momentum_P, J);
+        m_linear_momentum_P_norm =
+            sqrt(m_linear_momentum_P[0] * m_linear_momentum_P[0] +
+                 m_linear_momentum_P[1] * m_linear_momentum_P[1] +
+                 m_linear_momentum_P[2] * m_linear_momentum_P[2]);
         double J_norm = sqrt(J[0] * J[0] + J[1] * J[1] + J[2] * J[2]);
         m_mass = calculate_mass(m_area, J_norm);
         m_irreducible_mass = calculate_irreducible_mass(m_area);
 #elif CH_SPACEDIM == 2
-        m_mass = calculate_mass(m_area, 0.);
+        Tensor<1, double> P;
+        calculate_ah_quantities(m_area, P);
+        // assume that P_y = 0 in Cartoon code (if we calculate it it
+        // will not be 0, but that is because in the Cartoon code the
+        // AH is only half an horizon, so the P_y gets wrongly
+        // calculated)
+        m_linear_momentum_P_norm = P[0];
+        double J_norm = 0.;
+        m_mass = calculate_mass(m_area, J_norm);
 #endif
 
 #if CH_SPACEDIM == 3
@@ -365,26 +380,42 @@ void ApparentHorizon<SurfaceGeometry, AHFunction>::solve(double a_dt,
         m_spin_z_alt = calculate_spin_dimensionless(m_area);
 #endif
 
+#else // GR_SPACEDIM != 3
+        calculate_ah_quantities(m_area);
+#endif
+
         if (m_params.verbose > AHParams::NONE)
         {
-            pout() << "mass = " << m_mass << endl;
+#if GR_SPACEDIM == 3 // GR_SPACEDIM, not CH_SPACEDIM !!!
+            pout() << "mass = " << m_mass << std::endl;
 #if CH_SPACEDIM == 3
-            pout() << "spin = " << m_spin << endl;
-
+            pout() << "spin = " << m_spin << std::endl;
[…]
