+++ Examples/SingleBH/Main_SingleBH.cpp
@@ -25,6 +25,9 @@ int runGRChombo(int argc, char *argv[])
     GRParmParse pp(argc - 2, argv + 2, NULL, in_file);
     SimulationParameters sim_params(pp);
 
+    if (sim_params.just_check_params)
+        return 0;
+
     // The line below selects the problem that is simulated
     // (To simulate a different problem, define a new child of AMRLevel
     // and an associated LevelFactory)
