### 2021-08-18 #3564 @KAClough
Hi @rkurchin sorry for being slow - looking at the spreadsheet of reviewers I think the top ones would be shreyasbapat or sbozzolo, if they'd be willing to do it. If not I can take another look - the key elements are probably the C/C++ nature of the code, and an understanding of numerical relativity, although cosmological or hydrodynamical simulations obey similar equations so that would be sufficient too.
