### 2021-08-23 #3564 @rkurchin
👋 Hi there @shreyasbapat and @sbozzolo, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Please feel free to let me know if you have any questions about the process.
