### 2021-08-23 #3564 @Sbozzolo
Hi @rkurchin, this seems like a great submission for my expertise. (This would be my first review for JOSS.)
