### 2021-09-08 #3564 @rashti-alireza
Hi Rachel,
Sure.
Please let me know how I can access the code and the pertinent information.
Thanks.

*Kind regards,*
*Alireza Rashti.*

On Tue, Sep 7, 2021 at 6:27 PM Rachel Kurchin ***@***.***>
wrote:

> 👋 @rashti-alireza <https://github.com/rashti-alireza>, would you be
> willing to review this submission for JOSS? We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Please feel free to let me know if you have any questions about the
> process.
>
