### 2021-09-14 #3703 @Sbozzolo
Hi @KAClough. The paper is well written. I have two minor comments.

- It is often said that T_munu is the matter stress-energy tensor. This is not entirely accurate. For example, you can have electromagnetic fields.
- gamma_ij is never defined in the first bullet point in page 2.

Also, mostly for my own curiosity. I use the Einstein Toolkit, and it seems like that it has pretty much the same features, with exception of heavy use of templating. What are there reasons to use GRChombo as opposed to the Einstein Toolkit?
