### 2021-09-14 #3703 @KAClough
Thanks for reviewing @Sbozzolo :-) I will amend the draft for your two points, which I agree with.

It is true that we have many of the same features as ET - our main difference is that the adaptive mesh refinement is more flexible, so whereas in the ET with the Carpet driver you have a hierarchy of fixed size boxes that follow the compact objects, in our case the refined regions can be any shape or size. For compact objects the two approaches give very similar meshes and tbh I would say there is not much advantage in using GRChombo, but in the case of, for example, cosmic strings where you have a long extended object, this is (as far as I understand it) very difficult if not impossible with the Carpet mesh refinement driver. Actually I hear that ET is introducing more flexible AMR like that provided by Chombo (with AMRex instead), but I think this is still an ongoing project (led by Roland Haas).

Apart from this the use of templating is, as you say, one big feature, which improves the vectorisation of the RHS and also allows us to change elements like the matter content more flexibly. We don't use any code generation software like Kranc, so our users have to write things directly in code form - usually their goal is to do something new with the code, like modified gravity or exotic fields, so it is less designed to be a black box code and more one in which users get their hands dirty with coding. So again, although you can do binary BH mergers with it just fine, GRChombo is probably less straightforward than using the ET in this case. But if you want to do something with new initial conditions or fields, then the goal is that it should be more accommodated by the code structure in GRChombo than something like ET where some GR aspects are hard coded.

I should caveat that I have never used the ET, so most of what I say comes from chatting to other users. But I think in summary, GRChombo is meant to be targeted at non standard GR problems in fundamental physics, beyond the binary BH and neutron star cases (and as you will see we don't yet have fluids in the code).
