### 2021-09-14 #3703 @KAClough
Ah, thanks for that, we will fix the doxygen link.

For the templating, one example is that the CCZ4 evolution class with matter MatterCCZ4RHS is templated over a matter class, like a scalar or vector field. It is quite straightforward for a user to write their own matter class, by taking the ScalarField one and changing it, and then updating an example to use it just means replacing the template argument with their new matter type. We take the same approach in many places in the code, where we think a user might want to change the physics, we pass a template argument rather than hard code for a particular case.

It's true that Canuda is a good example of where ET has been used for similar kinds of fundamental physics applications. I think it has also been used for higher dimensional work in some cases as GRChombo has - by Uli Sperhake using his Lean code for example. I would say that GRChombo is built mainly to support these kind of use cases though, so it gets thought about a lot more at all levels of the code - in designing things like diagnostic quantities for example - whereas perhaps ET has a wider focus. But for sure you can do most of the same things with both codes, especially once AMReX is implemented in ET.
