### 2021-09-14 #3703 @Sbozzolo
Thanks!

I cannot find an overview of the features available/physics implemented. Does a such page exist? I think that it is really important to clearly present the capabilities of the code, so that people that have research projects in mind can see if GRChombo is the right tool to use or not. For instance, I couldn't clearly find if the code implements any quasi-isolated diagnostic for apparent horizons, or if I can evolve Einstein-Maxwell-Scalar with GRChombo.

If this page exists, it should be made easier to find.
