### 2021-09-15 #3703 @KAClough
We give a brief introduction to what the code can do in the "About GRChombo" page here:
https://github.com/GRChombo/GRChombo/wiki/About-GRChombo

and list tools here:
https://github.com/GRChombo/GRChombo/wiki/Existing-tools

But some of these pages are "under construction: and I agree it is a bit minimal - we can certainly update them to make them more clear. As it says in the About GRChombo page, we currently only keep the minimally coupled scalar field in the public version of the code, and not vector fields for the moment. The apparent horizon finder is also not yet in the public version of the code - although one exists in a branch, we want to make it more user friendly before adding it to the main code.
