### 2021-09-15 #3703 @rashti-alireza
Hi all,

@KAClough,
I appreciate the efforts that you and the others put to develop GRChombo and then made it open source to be used by the others as well ...

I am at the phase of installation, I have noticed a few minor things. Although they are mostly targeting Chombo but for the sake of completeness I have listed them as follows:

o. At "https://github.com/GRChombo/GRChombo/wiki/Compiling-Chombo" somewhere in the middle explained about HDF5
"If you want to use HDF5 (you probably do) ...".
The FALSE option was not explained for the variable `USE_HDF`.  I think explaining this option at the page is helpful for people who are not interested in HDF5 output and want to have a quick compilation for other purposes.
The other prerequisites are quit straightforward to install but to get HDF5 library right it might be a bit difficult in particular for
HPC machines.
Other than this, everything is explained very well regarding compilation and it is very good that there is a "Compilation FAQ" section for the common problems.

o. While compiling Chombo on my PC with the GNU compiler it gave a few warnings. I attached the pertinent files. No warning for GRChombo.

o. While compiling Chombo and GRChombo on Koko HPC with Intel compiler,  it also gave warnings.
The GRChombo warnings are probably benign but please be aware that they exist (at least for this version of compiler).
I attached them too.

o. Also I have a question: does it compile for windows operating system?

Thanks :)

[Make.defs.Koko_HPC.txt](https://github.com/openjournals/joss-reviews/files/7171770/Make.defs.Koko_HPC.txt)
[Make.defs.local.myPC.txt](https://github.com/openjournals/joss-reviews/files/7171771/Make.defs.local.myPC.txt)
[Koko_HPC_module_list.txt](https://github.com/openjournals/joss-reviews/files/7171772/Koko_HPC_module_list.txt)
[make_all_Chombo_myPC.txt](https://github.com/openjournals/joss-reviews/files/7171773/make_all_Chombo_myPC.txt)
[make_all_Chombo_Koko_HPC.txt](https://github.com/openjournals/joss-reviews/files/7171774/make_all_Chombo_Koko_HPC.txt)
[make_examples_GRChombo_Koko_HPC.txt](https://github.com/openjournals/joss-reviews/files/7171775/make_examples_GRChombo_Koko_HPC.txt)
