### 2021-09-16 #3703 @KAClough
It really is a question of how much work the user would need to do to recreate the results :-)

For vacuum spacetimes - ie BH binaries - everything is already in the public code including an adapted version of TwoPunctures for initial conditions. For any scalar field work (cosmology, ECOs, Scalar DM etc) again all of the ingredients are in the code already, but the user may need to amend the potential or initial conditions for the scalar to match the work since we don't provide the exact setups from the research projects in Examples. This is straightforward to do though.

For the cosmic string and higher dimensional/MG work, there are additional ingredients that the user would need that they could implement within the public code, but that are less straightforward to write from scratch. Ultimately all of these will end up in the public code - our goal is to add any new ingredients we use in our research to the public code within a year of the publication being accepted to a journal, but sometimes this timescale slips a bit as we are a fairly small team. The issue is that we don't want to put tools into the code in a state that isn't very readable or well commented, or where aspects are hardcoded that could be templated, for example, so often that requires a substantial rewrite and testing even though we have working code.

We are usually very happy to share our (working but ugly!) research codes with GRChombo users before they go into the public code though :-)
