### 2021-09-16 #3703 @mirenradia
Hi, @rashti-alireza, let me try and respond to some of your comments.

> o. At "https://github.com/GRChombo/GRChombo/wiki/Compiling-Chombo" somewhere in the middle explained about HDF5
> "If you want to use HDF5 (you probably do) ...".
> The FALSE option was not explained for the variable `USE_HDF`. I think explaining this option at the page is helpful for people who are not interested in HDF5 output and want to have a quick compilation for other purposes.

The reason this is not explained is because pretty much all numerical relativity users will want to build Chombo with HDF5 support (it is necessary in order to be able to use the checkpointing feature). In fact one of the only times I have set `USE_HDF=FALSE`, is for the github action that builds and runs the GRChombo tests with the Intel compiler and this is only because there is not an HDF5 package built with the Intel MPI library available in an Ubuntu software repository for the github runners.

> The other prerequisites are quit straightforward to install but to get HDF5 library right it might be a bit difficult in particular for
> HPC machines.

I agree that it would be a nontrivial package to build from source but I've never come across an HPC cluster that didn't have a working HDF5 module available.

> Other than this, everything is explained very well regarding compilation and it is very good that there is a "Compilation FAQ" section for the common problems.

Thank you for the nice feedback on these wiki pages! I appreciate it.

> o. While compiling Chombo on my PC with the GNU compiler it gave a few warnings. I attached the pertinent files. No warning for GRChombo.
>
> o. While compiling Chombo and GRChombo on Koko HPC with Intel compiler, it also gave warnings.
> The GRChombo warnings are probably benign but please be aware that they exist (at least for this version of compiler).
> I attached them too.

I'm aware of these warnings. They can be safely ignored.

> o. Also I have a question: does it compile for windows operating system?

I must admit that I have never tried but I highly doubt it. It should probably work under recent versions of the Windows Subsystem for Linux (WSL) though assuming things are set up correctly.
