### 2021-09-17 #3703 @rashti-alireza
Also in line 85, regarding "efficient strong scaling ...", Is the any plot, reference or similar for this?
