### 2021-09-19 #3703 @Sbozzolo
I am compiling Chombo on Expanse with the attached option file. The compilation goes well, but the tests fail. For the attached file, refluxEdgeTest fails with Error 3. Alternatively, I tried to set `OPT` to `FALSE` and enable `DEBUG`. In that case, other errors pop up, e.g.:

```
Running testBiCGStab for configuration 3d.Linux.64.mpiicpc.ifort.DEBUG.MPI.OPENMPCC ...
corrupted double-linked list

===================================================================================
=   BAD TERMINATION OF ONE OF YOUR APPLICATION PROCESSES
=   RANK 0 PID 125332 RUNNING AT exp-9-55
=   KILLED BY SIGNAL: 9 (Killed)
===================================================================================

===================================================================================
=   BAD TERMINATION OF ONE OF YOUR APPLICATION PROCESSES
=   RANK 1 PID 125333 RUNNING AT exp-9-55
=   KILLED BY SIGNAL: 6 (Aborted)
===================================================================================
... testBiCGStab finished with status 255
```

How do I debug this?

[expanse.txt](https://github.com/openjournals/joss-reviews/files/7192913/expanse.txt)
