### 2021-09-20 #3703 @KAClough
Hi @Sbozzolo is this just for doing 'make run' for the Chombo tests? Do the GRChombo tests run ok? If they do, then you can usually safely ignore the errors in the Chombo tests - it is not ideal but these don't seem to affect the code used by GRChombo. Usually you can get rid of them by trying different compiler combinations depending on what modules you have available on the system, but I'd not suggest spending time doing this as GRChombo should work fine anyway. This is noted in the compilation guide.
