### 2021-09-20 #3703 @KAClough
No problem, it is a bit of a long process to compile all the parts of the code. Hopefully not a lot worse than ET?! I've never tried compiling other NR codes, so I'm not sure how bad we are relatively speaking :-)
