### 2021-09-20 GRChombo/GRChombo#187 @Sbozzolo
[JOSS] Improvements to documentation and some missing information
Hello, I am [reviewing GRChombo for publication in the Journal of Open-Source Software (JOSS)](https://github.com/openjournals/joss-reviews/issues/3703). As part of the process, I am evaluating the quality of the documentation available. My overall impression is that the documentation in GRChombo contains good material, but it needs some polishing. Please, find below my comments. Please, keep in mind that the spirit of the comments are to help you improve GRChombo.

- Capabilities of the code are unclear

  The "about" page briefly mentioned what the code can do, but it is definitely not comprehensive enough. Suppose I have a research project in mind. There's not enough information for me to evaluate if I can use GRChombo for it.

  For example:
  * What physics modules are implemented in the public code?
  * Can I compute surface/volume ADM masses, and momenta?
  * Can I locate apparent horizons?
  * Can I compute the quasi-local properties of such horizons?
  * Can I change the order of the integration or finite-difference scheme?
  * Is checkpointing available?
  * Can I change parameters while a simulation is running, or at checkpoint?
  * What boundary conditions are available?
  * What initial data are available?
  * Can I perform Cauchy Characteristic Extraction of gravitational waves?
  * Are interpolation operators available?
  * ...

  These are just some examples. I got the answer to some of these questions, but it would be important to have a summary of what the code can do.

- Documentation has "work in progress" pages.

  Having incomplete pages is not a sign of good documentation and I think it is a red flag for new users. Importantly, the "structure of the code" page is incomplete. This page is important because it is the main way you can help new developers wrap their heads around GRChombo. If I were a developer, I would have to track how the code works by myself, which takes a lot of time an energy.

- The doxygen documentation has several broken links and does not seem useful

  It is unclear to me what purpose is the doxygen documentation serving. For instance, it seems that examples are intermixed with the main library and that there's no clear organization of the content. The documentation also says "We currently do not guarantee that this doxygen is kept completely up to date." As a user, I don't know what I have to expect from this documentation. Also, there are broken links, for example [here](https://grchombo.github.io/Doxygen/html/globals_eval.html#index_c).

- Postprocessing is not well documented.

  The documentation links to [this page](https://github.com/GRChombo/Postprocessing_tools/tree/main/YTAnalysisTools) for post-processing with yt. But there's no explanation on how to use these scripts. Similarly for [visit](https://github.com/GRChombo/Postprocessing_tools/tree/main/VisItTools). The included scripts have no comments. E.g., what is the difference between `PlotSlice.py` and `PlotSliceRemix.py`? How to use and modify these scripts?

- The README could give a better first-impression to new users

  What catches my eye when I open the README is "IMPORTANT: Updating examples for diagnostic variables". I have no idea what this means, but I think it is off-putting to new users. The README doesn't have any image and the information are really bare bone. JOSS [requires](https://joss.readthedocs.io/en/latest/review_criteria.html) that "A high-level overview of this documentation should be included in a README file (or equivalent)". The README should be updated to give an overview (that includes capabilities, dependencies, installation instructions) more extensively and clearly. Adding images will also improve the README.

- The relationship between "public GRChombo" and "private" is not clarified.

  The "about" of the GitHub repo says "Main public version of the GRChombo code", which implies the existence of a private version. Indeed, several published results depend on unpublished code. The development model of GRChombo should be clarified. Is GRChombo openly developed, or is it that the collaboration develops private modules, use it for a while, and than releases them to the public? If yes, is any promise made on timelines/disclosures/...?

- Copyright

  Who exactly owns the copyright? The header files say "GRChombo collaboration". What does this mean? Is there a contributor-license agreement that one needs to sign?

- Versions

  What versioning scheme does GRChombo follow? Is there any guarantee of backwards compatibility?

  JOSS requires the existence of a version, how will the wiki (e.g., installation instruction) change when you start introducing versions? Maybe you can start by tagging a beta release that will become official upon acceptance to publication in JOSS.

- Great pages

  I appreciated pages like "Coding conventions", "Updating the code", "Useful resources".

- How to obtain help and report bugs is missing

  I couldn't find information about how to report bugs and how to obtain help. Are GitHub issues the preferred way? What kind of information is needed in the bug report? Where should I ask my questions?

  If the page "contact us" was intended to serve this purpose, it should be expanded to clarify this.

- About contributing

  Do contributors have to write tests for their new features? How are PR merged (e.g., rebase vs merger commit)? What are the conventions for branch names?

- Configuration files for famous machines

  The Einstein Toolkit ships with simfactory. This tool is used, among the other things, to compile the code. Simfactory contains a database of several supercomputers, so that the code is compiled and executed in the most optimized way (e.g., compilation flags, module to load, environment variables to set, number of MPI processes to use per node, ...). Users of the Einstein Toolkit can submit their configuration files to simfactory, so that everyone in the community benefits from that. I think this is really useful, so I would suggest you do something similar and encourage users to submit their configuration files to compile and run GRChombo. Since you already have some, you could have a "contrib" folder (to make clear that some are "official", other are contributed).

- The organization of the wiki could be improved

  Is the documentation designed to be read as a book (from start to finish), or searched through as a traditional wiki? The way the wiki is currently setup seems to suggest the former. However, I think that this is a little bit intimidating and dispersive. For example, the table of content places at the same level: tutorials, usage pages, developer information, tips&tricks, other info. On the other hand, the "Getting started" page is much more friendly to new users than the wiki toc. So, my recommendation is to decrease the amount of information in the "home" page, keeping only a few links that are relevant to different audiences. You can also encourage users/developers to use the "search" feature to find what they look for.

  It would also be nice to have a "How do I?/FAQ" page. If video tutorials (e.g., for conferences/workshops) are available, they should be linked too.

- Required citations is missing or not easy to find

  What references am I supposed to cite when I use GRChombo?

These are some of the comments that came to my mind going through the documentation, but I haven't gone through each single page carefully.
