### 2021-09-20 GRChombo/GRChombo#187 @KAClough
Thanks @Sbozzolo for this really useful feedback. We will update the wiki for these suggestions over the next couple of weeks.
