### 2021-09-20 #3703 @KAClough
> Also in line 85, regarding "efficient strong scaling ...", Is the any plot, reference or similar for this?

@rashti-alireza sorry I missed this, yes a good reference would be the thesis of developer Markus Kunesch.

See https://www.repository.cam.ac.uk/handle/1810/283135, section 2.4
