### 2021-09-20 GRChombo/GRChombo#187 @KAClough
@rashti-alireza if you have any other comments on the wiki please do feel free to add them here too.

I've made a note that we could add some points on the compiler warnings and the use of HDF5 that you already raised.
