### 2021-09-20 GRChombo/GRChombo#187 @Sbozzolo
I forgot to add a point I was meaning to add. I think that it would be beneficial to have a page where you document how the documentation should look like. This for both users and developers. For users, it can provide a way to understand how the documentation is structured. For developers, this would be a "guiding principle" to maintain the documentation clean and updated.

Personally, I like this [mental framework](https://documentation.divio.com/) to organize my documentation. You can follow any framework you want, but I would recommend writing it down.
