### 2021-09-22 GRChombo/GRChombo#187 @Sbozzolo
* What are the recommendation for number of MPI processes vs OpenMP threads?
