### 2021-09-22 #3703 @Sbozzolo
`GRChombo` compiled fine and the examples run with no error.

When running tests, it might be nice to have a summary. E.g., something that prints something like "8 tests run: tests passed: [LIST], tests failed [LIST]". In `Compiling Chombo` you recommend using `grep` to obtain something like this. You don't do it in the "Compiling GRChombo" page. Maybe you can put together a simple `make` target to do this nicely?
