### 2021-09-28 #3703 @Sbozzolo
I cannot the binary BH example to run. I have this error:
```
[1632356692.166996] [exp-6-03:45110:0]         rcache.c:522  UCX  ERROR failed to insert region 0x1881ffac0 [0x22d64950..0x22d6df50]: Element already exists
[1632356692.167016] [exp-6-03:45110:0]         ucp_mm.c:116  UCX  ERROR failed to register address 0x22d64950 length 38400 on md[5]=mlx5_2: Element already exists
[1632356692.167019] [exp-6-03:45110:0]    ucp_request.c:268  UCX  ERROR failed to register user buffer datatype 0x8 address 0x22d64950 len 38400: Element already exists
```

Are there tests that are run with more than one MPI process?

The problem might be related to the MPI implementation I am using (Intel). I haven't debug this too deeply. Now I want to try with GCC and OpenMPI. In this, I realized that the documentation does not say anything about how to properly set Blas/Lapack.

Also, what is the expected output for the examples (ie, how do I know that I am doing things correctly)?
