### 2021-09-28 #3703 @KAClough
> I cannot the binary BH example to run. I have this error:
>
> ```
> [1632356692.166996] [exp-6-03:45110:0]         rcache.c:522  UCX  ERROR failed to insert region 0x1881ffac0 [0x22d64950..0x22d6df50]: Element already exists
> [1632356692.167016] [exp-6-03:45110:0]         ucp_mm.c:116  UCX  ERROR failed to register address 0x22d64950 length 38400 on md[5]=mlx5_2: Element already exists
> [1632356692.167019] [exp-6-03:45110:0]    ucp_request.c:268  UCX  ERROR failed to register user buffer datatype 0x8 address 0x22d64950 len 38400: Element already exists
> ```

Hmm, I have not seen this before - but if you are using intel 19+ it might help to try adding the -traditional flag - see point 9 of the FAQ (https://github.com/GRChombo/GRChombo/wiki/Compilation-FAQ). The error doesn't usually appear this way but it might be different on your system.

> Are there tests that are run with more than one MPI process?

You can run any of the individual tests with mpi by going into the individual folders and running the executables with mpirun (or equivalent). I think we deliberately didn't design the tests to run with mpi as we don't know want the user to accidentally overload the login nodes if they run the tests on them (as most of us do).

> The problem might be related to the MPI implementation I am using (Intel). When I run it in an interactive session, I get
>
> ```
> [exp-9-55:125919:0:126449] Caught signal 11 (Segmentation fault: invalid permissions for mapped object at address 0x15542a127478)
> [exp-9-55:125916:0:126524] Caught signal 11 (Segmentation fault: invalid permissions for mapped object at address 0x15542a1274d0)
> [exp-9-55:125939:0:126439] Caught signal 11 (Segmentation fault: invalid permissions for mapped object at address 0x15542a1278f0)
> [exp-9-55:125917:0:126503] Caught signal 11 (Segmentation fault: invalid permissions for mapped object at address 0x15542a127690)
> ```
>
> I haven't debug this too deeply. Now I want to try with GCC and OpenMPI. In this, I realized that the documentation does not say anything about how to properly set Blas/Lapack.

A lot of the systems we have used automatically link to these libraries in the compilation, but there are instructions on linking manually at the bottom of this wiki page:
https://github.com/GRChombo/GRChombo/wiki/Compiling-Chombo
And in point 2 of the Compilation FAQs.
There are also a number of example make files in https://github.com/GRChombo/GRChombo/tree/main/InstallNotes/MakeDefsLocalExamples
which give an idea of the types of flags that might be needed on specific systems and with specific compiler options.

> Also, what is the expected output for the examples (ie, how do I know that I am doing things correctly)?

As described here, checking you get the right puncture tracks and GW output for the binary example is a very good overall check that things are working correctly:
https://github.com/GRChombo/GRChombo/wiki/Running-the-BinaryBH-example
