### 2021-09-28 GRChombo/GRChombo#187 @KAClough
Moving @Sbozzolo comment from the review here as it is related to updating the code for JOSS paper:

>> When running tests, it might be nice to have a summary. E.g., something that prints something like "8 tests run: tests passed: [LIST], tests failed [LIST]". In Compiling Chombo you recommend using grep to obtain something like this. You don't do it in the "Compiling GRChombo" page. Maybe you can put together a simple make target to do this nicely?
