### 2021-09-28 #3703 @Sbozzolo
> A lot of the systems we have used automatically link to these libraries in the compilation, but there are instructions on linking manually at the bottom of this wiki page:
> https://github.com/GRChombo/GRChombo/wiki/Compiling-Chombo

The machine I am using did not do that. Looking at the examples, it seems that MKL is always used. When GNU compilers are used, you need to specify inc and lib paths for MKL (or any other library you want to use).

I find the information and the setup a little bit inconsistent. For example, there are variables `HDF5INCFLAGS`, so I was expecting to see something similar for `LAPACK` and `Blas`. Also, `syslibflags` seems to be for the linking stage, but one needs to specify the include directory as well (with `-I/xxxx/`). This is not said in either the page nor the FAQ. I am personally familiar on how all these things work and I have no problems understanding what is going on, but I think this might be a source of confusion to those users that are trying to compile on a new system. Of course, the examples help a lot, but they are probably helpful only to those that can understand them.

> As described here, checking you get the right puncture tracks and GW output for the binary example is a very good overall check that things are working correctly:
> https://github.com/GRChombo/GRChombo/wiki/Running-the-BinaryBH-example

Sorry, I should have been more specific. I have never run GRChombo before and I have no idea on what to expect from the STDOUT and STDERR. In my case, in the `out` file I see these lines:
```
 number_procs = 64
 threads = 4
 simd width (doubles) = 2
```
And nothing else. What files should I find?
