### 2021-09-28 #3703 @Sbozzolo
> You can run any of the individual tests with mpi by going into the individual folders and running the executables with mpirun (or equivalent). I think we deliberately didn't design the tests to run with mpi as we don't know want the user to accidentally overload the login nodes if they run the tests on them (as most of us do).

When I run the tests with `mpirun`, is the code going through all the associated non-trivial steps? E.g., domain decomposition across MPI ranks?

Also, is there a guide on how to write tests?
