### 2021-09-28 #3703 @mirenradia
> I cannot the binary BH example to run. I have this error:
>
> ```
> [1632356692.166996] [exp-6-03:45110:0]         rcache.c:522  UCX  ERROR failed to insert region 0x1881ffac0 [0x22d64950..0x22d6df50]: Element already exists
> [1632356692.167016] [exp-6-03:45110:0]         ucp_mm.c:116  UCX  ERROR failed to register address 0x22d64950 length 38400 on md[5]=mlx5_2: Element already exists
> [1632356692.167019] [exp-6-03:45110:0]    ucp_request.c:268  UCX  ERROR failed to register user buffer datatype 0x8 address 0x22d64950 len 38400: Element already exists
> ```
>

What cluster are you trying to run on? I have never seen this problem before either and you might need to speak to cluster support in order to resolve the problem. However, it's possible someone else in the GRChombo collaboration has used the system before so we may be able to help out.

> Are there tests that are run with more than one MPI process?
> When I run the tests with `mpirun`, is the code going through all the associated non-trivial steps? E.g., domain decomposition across MPI ranks?

If you have built Chombo with `MPI=TRUE`, all tests are called with `mpirun -np 2` by default (even if the individual tests do not call any MPI functions in which case the test just runs twice in separate processes). One test that uses MPI non-trivially and performs decomposition of the domain across MPI ranks is the AMRInterpolator test.

> Also, is there a guide on how to write tests?

No. This is quite problem-specific and is up to the reviewer of the PR.

>
> The problem might be related to the MPI implementation I am using (Intel). When I run it in an interactive session, I get
>
> ```
> [exp-9-55:125919:0:126449] Caught signal 11 (Segmentation fault: invalid permissions for mapped object at address 0x15542a127478)
> [exp-9-55:125916:0:126524] Caught signal 11 (Segmentation fault: invalid permissions for mapped object at address 0x15542a1274d0)
> [exp-9-55:125939:0:126439] Caught signal 11 (Segmentation fault: invalid permissions for mapped object at address 0x15542a1278f0)
> [exp-9-55:125917:0:126503] Caught signal 11 (Segmentation fault: invalid permissions for mapped object at address 0x15542a127690)
> ```

Is this also the BinaryBH example? Without further debugging, I'm not sure I can help much further.

> I haven't debug this too deeply. Now I want to try with GCC and OpenMPI. In this, I realized that the documentation does not say anything about how to properly set Blas/Lapack.

This depends very much on the specific BLAS/LAPACK implementations and how they have been installed. For example, on my own Ubuntu 20.04 laptop, with the `liblapack-dev` and `libblas-dev` packages installed, it is sufficient to add `-lblas -llapack` to the `syslibflags` variable in the `Make.defs.local` (see point 2 in the [Compilation FAQ](https://github.com/GRChombo/GRChombo/wiki/Compilation-FAQ)). If you are on a cluster with Intel MKL installed, you can use the [Intel MKL Link Line Advisor](https://software.intel.com/content/www/us/en/develop/tools/oneapi/components/onemkl/link-line-advisor.html) assuming the MKL module is loaded (and thus the `MKLROOT` environment variable has been correctly set). We should add a link to this tool on the compilation FAQ.

>I find the information and the setup a little bit inconsistent. For example, there are variables HDF5INCFLAGS, so I was expecting to see something similar for LAPACK and Blas. Also, syslibflags seems to be for the linking stage, but one needs to specify the include directory as well (with -I/xxxx/). This is not said in either the page nor the FAQ. I am personally familiar on how all these things work and I have no problems understanding what is going on, but I think this might be a source of confusion to those users that are trying to compile on a new system. Of course, the examples help a lot, but they are probably helpful only to those that can understand them.

I agree that the build system is inconsistent and I'm personally not a fan of it but unfortunately we have inherited this from the Chombo library (if I were designing it from scratch, I'd probably use CMake). If the modules on your cluster correctly set the `CPATH` environment variable, then there shouldn't be a need to add extra include flags.

>Sorry, I should have been more specific. I have never run GRChombo before and I have no idea on what to expect from the STDOUT and STDERR. In my case, in the out file I see these lines:
>``` number_procs = 64
> threads = 4
> simd width (doubles) = 2```
> And nothing else. What files should I find?

As per the [Running examples wiki page](https://github.com/GRChombo/GRChombo/wiki/Running-examples), the code writes "STDOUT"-style output to a `pout` file from each process (e.g. `pout.0` from rank 0, `pout.1` from rank 1, etc.). These can be found in the relative directory specified by the [`pout_subpath` parameter](https://github.com/GRChombo/GRChombo/blob/4baece8164eef6fae5e1b29d929ef324d9e8081b/Examples/BinaryBH/params.txt#L28). Similarly, you can find HDF5 file output in the relative directory specified by the [`hdf5_subpath` parameter](https://github.com/GRChombo/GRChombo/blob/4baece8164eef6fae5e1b29d929ef324d9e8081b/Examples/BinaryBH/params.txt#L27) and GW extraction data in the relative directory specified by the [`data_subpath` parameter](https://github.com/GRChombo/GRChombo/blob/4baece8164eef6fae5e1b29d929ef324d9e8081b/Examples/BinaryBH/params.txt#L29).
