### 2021-09-28 #3703 @Sbozzolo
> What cluster are you trying to run on? I have never seen this problem before either and you might need to speak to cluster > support in order to resolve the problem.

I am running on Expanse, at SDSC. The reason I wanted to switch to GCC+OpenMPI is because it doesn't use ucx (on Expanse), so it might fix the problem.

> If you have built Chombo with MPI=TRUE, all tests are called with mpirun -np 2 by default (even if the individual tests do not call any MPI functions in which case the test just runs twice in separate processes). One test that uses MPI non-trivially and performs decomposition of the domain across MPI ranks is the AMRInterpolator test.

Since GRChombo depends critically on MPI working as expected, it would be nice to have some more tests in that department. I am especially concerned with the lower level parts, so if Chombo has good MPI tests, that would probably be fine too.

> No. This is quite problem-specific and is up to the reviewer of the PR.

It would be nice to have at least a basic description of how testing works. For example, I was expecting to find a data file with the reference value of the output for each test that is compared with the actual output. When I open a test, eg. BSSNMatterTest.cpp, I don't see what is being compared to what.

> Is this also the BinaryBH example?

Correct

> This depends very much on the specific BLAS/LAPACK implementations and how they have been installed. For example, on my own Ubuntu 20.04 laptop, with the liblapack-dev and libblas-dev packages installed, it is sufficient to add -lblas -llapack to the syslibflags variable in the Make.defs.local (see point 2 in the Compilation FAQ).

This is because on your laptop all the libraries are installed in default locations (/usr/lib, /lib64, etc). This is not true on clusters. The module system should take care of this, but this doesn't always work.

I am not asking to overhaul the build system, but I think that it is important that everything is properly documented and accessible.

> As per the Running examples wiki page, the code writes "STDOUT"-style output to a pout file from each process (e.g. pout.0 from rank 0, pout.1 from rank 1, etc.). These can be found in the relative directory specified by the pout_subpath parameter. Similarly, you can find HDF5 file output in the relative directory specified by the hdf5_subpath parameter and GW extraction data in the relative directory specified by the data_subpath parameter.

Okay, I was looking at the output as SLRUM captured it, and I didn't see anything interesting there. Should I `tail` `pout.0` to get a sense of what is going on?
