### 2021-09-28 #3703 @mirenradia
> I am running on Expanse, at SDSC. The reason I wanted to switch to GCC+OpenMPI is because it doesn't use ucx (on Expanse), so it might fix the problem.
>

Hmm, I have heard several complaints about Expanse and the one GRChombo user who tried it (@kazewong) gave up and switched to a different cluster he had access to due to unreliable performance. Do you have access to a different (perhaps more-established) cluster to test on?

> Since GRChombo depends critically on MPI working as expected, it would be nice to have some more tests in that department. I am especially concerned with the lower level parts, so if Chombo has good MPI tests, that would probably be fine too.

These "lower level parts" are part of the Chombo library and are thus tested there. Having said that, a few tests that are irrelevant to GRChombo often fail (e.g. the refluxEdgeTest you mentioned previously) so we usually advise users to ignore these tests when running on a new cluster and rely on the GRChombo tests.

> It would be nice to have at least a basic description of how testing works. For example, I was expecting to find a data file with the reference value of the output for each test that is compared with the actual output. When I open a test, eg. BSSNMatterTest.cpp, I don't see what is being compared to what.

I think the nature of a test varies quite a lot depending on the feature being tested. Actually in the [CCZ4 Geometry Unit Tests](https://github.com/GRChombo/GRChombo/tree/main/Tests/CCZ4GeometryUnitTests), you can find a Mathematica notebook which generates values that are used to test the calculation of the Ricci tensor. In the case of the [BSSN Matter Test](https://github.com/GRChombo/GRChombo/tree/main/Tests/BSSNMatterTest), this compares the values outputted by the current version of the code with an [old Fortran version](https://github.com/GRChombo/GRChombo/blob/main/Tests/BSSNMatterTest/GRBSSNChomboF.ChF) that is known to produce correct results.

> This is because on your laptop all the libraries are installed in default locations (/usr/lib, /lib64, etc). This is not true on clusters. The module system should take care of this, but this doesn't always work.
>
> I am not asking to overhaul the build system, but I think that it is important that everything is properly documented and accessible.

Sorry, I've just remembered that it shouldn't actually be necessary to specify the include paths for BLAS/LAPACK as no header files are used from them

> Okay, I was looking at the output as SLRUM captured it, and I didn't see anything interesting there. Should I `tail` `pout.0` to get a sense of what is going on?

Yes I typically run `tail -f pout.0` to check the progress of a simulation. Bear in mind that not much information is given in the case the `verbosity` parameter is set to 0.
