### 2021-09-30 GRChombo/GRChombo#187 @Sbozzolo
Thanks @mirenradia. Let me point out that my questions were not intended to be answered directly, but they were prompts for you to add and improve the documentation. Users should be able to find this information in the documentation.

> I agree that a slightly more detailed overview of what is currently available is a good idea. However, I think it might be difficult to maintain an up-to-date list of every single feature that is available given the small size of our collaboration (relative to the ETK collaboration). I note that JOSS specifies the "core functionality" be understood.

I argue that some of the questions are asked are definitely "core functionality" (e.g., checkpointing). And, personally, I think that being a small team makes things easier. Once you have the initial list of features available, it is not going to change dramatically and/or over fast timescales, so you can more easily keep it updated. I believe that having a clear list of features is a big service to new users, and I highly encourage this, regardless of what are the standards in JOSS. For example, the Einstein Toolkit does not have that, so, in the past, I had to write codes that already existed because I was unaware of their existence.
