### 2021-10-05 ccd20e4
Add github action to build Doxygen documentation
