### 2021-10-06 #3703 @rashti-alireza
@KAClough,
I tried to run BBH example with the parameter file provided in the folder. It seems the BHs getting stuck in a quarter of an orbit.
I attached the parameter file I used and puncture.dat output.

[bbh_GRChombo_cheap.txt](https://github.com/openjournals/joss-reviews/files/7290317/bbh_GRChombo_cheap.txt)
[punctures.dat.txt](https://github.com/openjournals/joss-reviews/files/7290318/punctures.dat.txt)
