### 2021-10-06 7e7d992
Update joss paper for reviewer comments
