+++ paper.bib
@@ -1,3 +1,12 @@
+@book{kreiss1973methods,
+  title={Methods for the Approximate Solution of Time Dependent Problems},
+  author={Kreiss, H.O. and Oliger, J. and Global Atmospheric Research Programme. Joint Organizing Committee and International Council of Scientific Unions and World Meteorological Organization},
+  series={GARP publications series},
+  url={https://books.google.co.uk/books?id=OxMZAQAAIAAJ},
+  year={1973},
+  publisher={International Council of Scientific Unions, World Meteorological Organization}
+}
+
 @article{Abbott:2016blz,
       author         = "Abbott, B. P. and others",
       title          = "{Observation of Gravitational Waves from a Binary Black
