### 2021-10-06 27e8010
Replace local with GitHub URLs in Doxygen docs
