### 2021-10-06 #3703 @rashti-alireza
> > @KAClough, I tried to run BBH example with the parameter file provided in the folder. It seems the BHs getting stuck in a quarter of an orbit. I attached the parameter file I used and puncture.dat output.
> > [bbh_GRChombo_cheap.txt](https://github.com/openjournals/joss-reviews/files/7290317/bbh_GRChombo_cheap.txt) [punctures.dat.txt](https://github.com/openjournals/joss-reviews/files/7290318/punctures.dat.txt)
>
> Hi @rashti-alireza, unfortunately the "cheap" parameter files exist merely to test the code does not crash (e.g. after modification) on a personal computer and the results they produce will not be physically accurate or meaningful due to the very low resolutions. If you wish to perform a simulation with a reasonable level of accuracy, you will probably need to use an HPC cluster and use a parameter file closer to [params.txt](https://github.com/GRChombo/GRChombo/blob/main/Examples/BinaryBH/params.txt).

OK, Thanks. Do you have any other parameter file example like head on collision?
