### 2021-10-06 476a4a0
Modify action to create PR to update Doxygen docs
