### 2021-10-07 b6209f0
Change Doxygen action trigger
Now it will only run when new commits are pushed to main.
