### 2021-10-07 b552736
Make Doxygen title page refer to wiki instructions
The instructions on the wiki are more comprehensive than that in the
Doxygen folder.
