### 2021-10-07 a41d88a
Add GitHub action to update Doxygen documentation (#195)
* Add github action to build Doxygen documentation

* Replace local with GitHub URLs in Doxygen docs

* Checkout GRChombo under GRChombo relative path

* Modify action to create PR to update Doxygen docs

* Refine automatic PR generated by doxygen action

* Fix assignment of Doxygen PR reviewer

* Change Doxygen action trigger

Now it will only run when new commits are pushed to main.

* Make Doxygen title page refer to wiki instructions

The instructions on the wiki are more comprehensive than that in the
Doxygen folder.
