### 2021-10-29 dd53963
Add pre-commit configuration file (#198)
This uses https://pre-commit.com/ and currently just runs the
clang-format hook.
