+++ .pre-commit-config.yaml
@@ -1,0 +1,5 @@
+repos:
+-   repo: https://github.com/pre-commit/mirrors-clang-format
+    rev: v10.0.1  
+    hooks:
+    -   id: clang-format
