### 2021-10-31 #3703 @mirenradia
>This is the Make.defs.local
>
>```
>CXX=g++ -I/opt/ohpc/pub/compiler/intel_2020_u4/compilers_and_libraries/linux/mkl/include -O3 -march=native
>FC=gfortran -I/opt/ohpc/pub/compiler/intel_2020_u4/compilers_and_libraries/linux/mkl/include -O3 -march=native
>MPICXX=mpicxx -I/opt/ohpc/pub/compiler/intel_2020_u4/compilers_and_libraries/linux/mkl/include -O3 -march=native
>PRECISION=DOUBLE
>OPENMPCC=TRUE
>MPI=TRUE
>DIM=3
>OPT=HIGH
>DEBUG=FALSE
>
>USE_HDF          = TRUE
>HDFINCFLAGS      = -I/opt/ohpc/pub/libs/gnu8/hdf5/1.10.5/include
>HDFLIBFLAGS      = -L/opt/ohpc/pub/libs/gnu8/hdf5/1.10.5/lib -lhdf5 -lz
>HDFMPIINCFLAGS   = -I/opt/ohpc/pub/apps/phdf5/1.12.0/include
>HDFMPILIBFLAGS   = -L/opt/ohpc/pub/apps/phdf5/1.12.0/lib -lhdf5 -lz
>syslibflags      = -L/opt/ohpc/pub/compiler/intel_2020_u4/compilers_and_libraries/linux/mkl/lib/intel64 -Wl,--no-as-needed -lmkl_intel_lp64 -lmkl_sequential -lmkl_core -lpthread -lm -ldl
>```

As I suspected nothing seems obviously wrong so let's see if we can get the Intel one working instead.

> I tried compiling with the intel compilers, but I couldn't get it to compile `GRChombo`. I tried with Intel 2020 and 2021. `Chombo` compiles fine.
>
> I tried to fiddle around and experiment with various compilation flags. This is a makefile I have just written to reproduce the error:
>
> ```
> CXX=icpc -O3 -xHost -mkl=sequential -std=c++14 -qopenmp
> FC=ifort -O3 -xHost -mkl=sequential -qopenmp
> MPICXX=mpiicpc -O3 -xHost -qopenmp -std=c++14 -mkl=sequential
> PRECISION=DOUBLE
> OPENMPCC=TRUE
> MPI=TRUE
> DIM=3
> OPT=HIGH
> DEBUG=FALSE
>
> USE_HDF          = TRUE
> HDFINCFLAGS      = -I/opt/ohpc/pub/apps/hdf5-intel/1.12.0/include
> HDFLIBFLAGS      = -L/opt/ohpc/pub/apps/hdf5-intel/1.12.0/lib -lhdf5 -lz
> HDFMPIINCFLAGS   = -I/opt/ohpc/pub/apps/phdf5-intel/1.12.0/include
> HDFMPILIBFLAGS   = -L/opt/ohpc/pub/apps/phdf5-intel/1.12.0/lib -lhdf5 -lz
> ```
>
> This is the error when I compile the binary black hole example: [err.txt](https://github.com/openjournals/joss-reviews/files/7443686/err.txt)
>
> I read the FAQs and triple-checked that I am using `-std=c++14`.

Since the Intel compiler uses the standard library header files from GCC, what version of GCC is available to it? If you haven't got a GCC module loaded and it's using the system one, this might be too old to have all the required standard library headers which would explain your compilation error. If this is indeed the issue, I'd recommend trying to load a newer one.
