### 2021-11-01 #3703 @Sbozzolo
> Since the Intel compiler uses the standard library header files from GCC, what version of GCC is available to it? If you haven't got a GCC module loaded and it's using the system one, this might be too old to have all the required standard library headers which would explain your compilation error. If this is indeed the issue, I'd recommend trying to load a newer one.

This worked! But it didn't solve the problem: after 40 minutes of running, the job hanged.

@rashti-alireza did you reproduce the binary black hole example?
