### 2021-11-02 #3703 @mirenradia
> This worked!

Great, we should add this to the FAQ page on the wiki.

> But it didn't solve the problem: after 40 minutes of running, the job hanged.

This is a bit of a mystery to me. I presume there's still no output to stderr? Looking at the last few lines of `pout.0`, does the output end at the same place? This suggestion is a bit of a shot in the dark but can you try running with the parameter `verbosity = 0` (I'm wondering if the buffer is full from too much output)?
