### 2021-11-02 #3703 @Sbozzolo
> > This worked!
>
> Great, we should add this to the FAQ page on the wiki.
>
> > But it didn't solve the problem: after 40 minutes of running, the job hanged.
>
> This is a bit of a mystery to me. I presume there's still no output to stderr? Looking at the last few lines of `pout.0`, does the output end at the same place? This suggestion is a bit of a shot in the dark but can you try running with the parameter `verbosity = 0` (I'm wondering if the buffer is full from too much output)?

There is no stderr or stdout. The last few lines of the `.out` file are (from 12120):
```
GRAMRLevel constructor
GRAMRLevel::define 7
GRAMRLevel constructor
GRAMRLevel constructor
GRAMRLevel constructor
GRAMRLevel::define 7
GRAMRLevel::define 7
GRAMRLevel::define 7
GRAMRLevel constructor
GRAMRLevel::define 7
GRAMRLevel constructor
GRAMRLevel::define 7
```
The last few lines of the `pout.0` file are (from lines 231755):
```
GRAMRLevel::computeDt 7
GRAMRLevel::postTimeStep 7
GRAMRLevel::postTimeStep 7 finished
GRAMRLevel::advance level 7 at time 176.016 (286.115 M/hr). Boxes on this rank: 11 / 1340
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::computeDt 7
GRAMRLevel::postTimeStep 7
GRAMRLevel::postTimeStep 7 finished
GRAMRLevel::postTimeStep 6
GRAMRLevel::postTimeStep 6 finished
GRAMRLevel::advance level 6 at time 176.031 (286.075 M/hr). Boxes on this rank: 4 / 484
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::computeDt 6
GRAMRLevel::advance level 7 at time 176.031 (286.028 M/hr). Boxes on this rank: 11 / 1340
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::computeDt 7
GRAMRLevel::postTimeStep 7
GRAMRLevel::postTimeStep 7 finished
GRAMRLevel::advance level 7 at time 176.047 (285.993 M/hr). Boxes on this rank: 11 / 1340
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::evalRHS
GRAMRLevel::computeDt 7
GRAMRLevel::postTimeStep 7
GRAMRLevel::postTimeStep 7 finished
GRAMRLevel::postTimeStep 6
GRAMRLevel::postTimeStep 6 finished
GRAMRLevel::postTimeStep 5
GRAMRLevel::fillAllEvolutionGhosts
GRAMRLevel::fillAllEvolutionGhosts
```

I am trying with no verbosity.
