### 2021-11-02 #3703 @mirenradia
Hmm, it seems the one built with the Intel compiler manages to get quite a bit further than the GCC one.

Another thing that confuses me is why you have output in your `.out` file that should be in the `pout` files. Looking at the [Chombo source code](https://github.com/GRChombo/Chombo/blob/main/lib/src/BaseTools/parstream.cpp#L123-L158), this should only happen if it fails to open the `pout` file or if `MPI_Initialized` is erroneously reporting that `MPI_Init` hasn't been called yet.
