### 2021-11-02 #3703 @Sbozzolo
With `verbosity = 0` it hangs exactly at the same time.
