### 2021-11-02 #3703 @mirenradia
> With `verbosity = 0` it hangs exactly at the same time.

Some more things to try
* Does the time when it hang changes if there is only 1 OpenMP thread (keep the number of ranks/node constant so that memory usage doesn't change much)?
* If you recompile with `OPT = TRUE` (rather than `HIGH` in your Make.defs.local), this will enable assertions. Try rerunning with this (it will probably significantly slow down your simulation) to see if you get a message about a failed assertion.
