### 2021-11-02 #3703 @mirenradia
Also, it would be worth ensuring your MPI launch command correctly distributes the MPI ranks over the node. If using Intel MPI, you can use the `-ppn` flag so your launch line would be something like
```
mpirun -np ${SLURM_NTASKS} -ppn ${SLURM_NTASKS_PER_NODE}  ./<exec>.ex params.txt
```
In an ideal world, it should be able to automatically determine and set this from the SLURM job but this is not always the case.
