### 2021-11-04 0b74868
Fix Weyl Imaginary part extraction bug
