### 2021-11-07 #3703 @Sbozzolo
`OPT = TRUE` shows that there's a nan: (from the error file)
```
Values have become nan. !!!
```

I tried removing every single optimization, and compiling in debug mode.
Now I get a segmentation fault:
```
srun: error: i0n2: task 1: Segmentation fault (core dumped)
```
This happens quite quickly, `pout` files end with:
```
GRAMRLevel::advance level 0 at time 0 (0 M/hr). Boxes on this rank: 1 / 32
GRAMRLevel::advance level 1 at time 0 (0 M/hr). Boxes on this rank: 1 / 83
```
