### 2021-11-08 #3703 @Sbozzolo
I forgot to mention that the NaN happens also with 1 OpenMP thread.

On Mon, Nov 8, 2021 at 11:37 AM Miren Radia ***@***.***>
wrote:

> OPT = TRUE shows that there's a nan: (from the error file)
>
> Values have become nan. !!!
>
> I tried removing every single optimization, and compiling in debug mode.
> Now I get a segmentation fault:
>
> srun: error: i0n2: task 1: Segmentation fault (core dumped)
>
> This happens quite quickly, pout files end with:
>
> GRAMRLevel::advance level 0 at time 0 (0 M/hr). Boxes on this rank: 1 / 32
> GRAMRLevel::advance level 1 at time 0 (0 M/hr). Boxes on this rank: 1 / 83
>
> @KAClough <https://github.com/KAClough> and I have found it difficult to
> reproduce this new problem and my suspicion is that it is unrelated to your
> hanging one. Would you be able to try the
> enhancement/disable_omp_time_interpolator Chombo branch
> <https://github.com/GRChombo/Chombo/tree/enhancement/disable_omp_time_interpolator>
> and see if that solves your problem? Alternatively, you can try running
> with 1 OpenMP thread.
>
