### 2021-11-08 #3703 @Sbozzolo
Also, I didn't give enough information, but the two problems are connected because the NaN happens at the same time (~176 M) as when I start to see the processes hanging.
