### 2021-11-08 #3703 @Sbozzolo
> Oh I think I got confused by your previous message and thought the NaN you saw was off the first timestep like the segfault you mentioned. So in all other cases, you get a NaN at exactly the same time (for some reason I thought this differed between GCC+OpenMPI and the Intel stack)? Would you be able to share the current parameter file you are using (if you made any modifications following @KAClough's suggestions to help with memory usage earlier) so that I can try and reproduce it?

Here it is!
[params.txt](https://github.com/openjournals/joss-reviews/files/7500772/params.txt)
