### 2021-11-09 #3703 @mirenradia
> Here it is! [params.txt](https://github.com/openjournals/joss-reviews/files/7500772/params.txt)

I can confirm that I also get NaNs at around t=176. However, this is not surprising as the resolution is way too low in order to resolve this simulation with any kind of accuracy. Looking at the puncture trajectory, we can see that the BHs have already merged when they shouldn't do so (for these binary parameters) until well after t=2000M (cf. the plot at the end of [this wiki page](https://github.com/GRChombo/GRChombo/wiki/Running-the-BinaryBH-example))

When @KAClough made the suggestion to reduce `max_level` from `9` to `7`, I think she was under the impression you just wanted to quickly check the code evolved for a bit. If you want to test this particular simulation with any kind of accuracy, you will need to set `max_level` to `9` again (and probably provide more computational resources).
