### 2021-11-10 #3703 @Sbozzolo
> > Here it is! [params.txt](https://github.com/openjournals/joss-reviews/files/7500772/params.txt)
>
> I can confirm that I also get NaNs at around t=176. However, this is not surprising as the resolution is way too low in order to resolve this simulation with any kind of accuracy. Looking at the puncture trajectory, we can see that the BHs have already merged when they shouldn't do so (for these binary parameters) until well after t=2000M (cf. the plot at the end of [this wiki page](https://github.com/GRChombo/GRChombo/wiki/Running-the-BinaryBH-example))
>
> When @KAClough made the suggestion to reduce `max_level` from `9` to `7`, I think she was under the impression you just wanted to quickly check the code evolved for a bit. If you want to test this particular simulation with any kind of accuracy, you will need to set `max_level` to `9` again (and probably provide more computational resources).

Sweet! Indeed, I wanted to check that the code evolved for a bit, and that configuration is not ideal for that. Now, I evolved for one orbit at full resolution facing no problems. Some comments:
* It is not great that the code hangs indefinitely when a NaN is encountered and `OPT = HIGH`.
* Are users going to receive the level of support I received in figuring this out? If yes, what is the correct avenue to ask for support? (This has to be clearly stated somewhere in the documentation)
* Postprocessing of the simulation data seems a little bit limited. I don't see any clear information on the structure of the HDF5 output; the Python tools provided by GRChombo are not documented. Note that postprocessing is not only visualizing (for example, postprocessing might include doing convergence tests on specific grid functions).
* Is it possible to change parameters between checkpoints?
* Is it possible to control the simulation while it is running (e.g, by writing in files that are watched by the code?)

The point of GRChombo is not to run the examples, but to build new unique simulations. I cannot test that, so I am going to consider myself satisfied with what I did and cross the code-related items. However, as we have seen, the process was not a cakewalk. So, to ensure the success and longevity of the project, my recommendation is to double down on the documentation and user support. Without these, it would very difficult for someone outside of the GRChombo sphere of influence to use the code and achieve their scientific goals (which, I think, is a good test for fitness for publication in JOSS: "If I give this code and documentation to someone, can they figure out on their own how to use it in a meaningful way?"). You are already in good shape with documentation (but see GitHub issue). I would also recommend adding some guidelines/rules for all the new code that is added to ensure that all the new code is documented in an impeccable way. As for user support, at the moment it seems to me that one has to send an email to grchombo@gmail.com. I think it would be beneficial to the community if all the support discussions were public (so that you avoid answering the same questions multiple times). So, if you want to offer support via email, maybe you can consider a mailing list with searchable archives. Alternatively, you can consider other options (e.g., GitHub discussions). Let me point out once again that these comments are here for you to have a outside point of view of on how to improve your project.

Once you address at least the more important points in https://github.com/GRChombo/GRChombo/issues/187, I will recommend this work to publication in JOSS.
