### 2021-11-10 #3703 @KAClough
> * Are users going to receive the level of support I received in figuring this out? If yes, what is the correct avenue to ask for support? (This has to be clearly stated somewhere in the documentation)

At some level yes, we often do help with debugging issues. There is a contact email on the wiki and users can file issues but I will highlight these options when I update the wiki.

> * Postprocessing of the simulation data seems a little bit limited. I don't see any clear information on the structure of the HDF5 output; the Python tools provided by GRChombo are not documented. Note that postprocessing is not only visualizing (for example, postprocessing might include doing convergence tests on specific grid functions).

Yes, we plan to update some of the postprocessing parts of the wiki. We do expect most people to write their own analysis tools depending on the problem, but there are a few examples they can start from which we could introduce more clearly.

> * Is it possible to change parameters between checkpoints?

Yes, for some quantities but not others. It is a good point and I think we should update the params guide in the wiki to highlight which ones can be amended (or at least how to find this out from looking at the code).

> * Is it possible to control the simulation while it is running (e.g, by writing in files that are watched by the code?)

Not for the moment - I think @mirenradia is implementing something which will allow stopping and restarting more easily mid simulation, but it is not yet in the public code.

> The point of GRChombo is not to run the examples, but to build new unique simulations. I cannot test that, so I am going to consider myself satisfied with what I did and cross the code-related items. However, as we have seen, the process was not a cakewalk. So, to ensure the success and longevity of the project, my recommendation is to double down on the documentation and user support. Without these, it would very difficult for someone outside of the GRChombo sphere of influence to use the code and achieve their scientific goals (which, I think, is a good test for fitness for publication in JOSS: "If I give this code and documentation to someone, can they figure out on their own how to use it in a meaningful way?"). You are already in good shape with documentation (but see GitHub issue). I would also recommend adding some guidelines/rules for all the new code that is added to ensure that all the new code is documented in an impeccable way. As for user support, at the moment it seems to me that one has to send an email to [grchombo@gmail.com](mailto:grchombo@gmail.com). I think it would be beneficial to the community if all the support discussions were public (so that you avoid answering the same questions multiple times). So, if you want to offer support via email, maybe you can consider a mailing list with searchable archives. Alternatively, you can consider other options (e.g., GitHub discussions). Let me point out once again that these comments are here for you to have a outside point of view of on how to improve your project.
>
> Once you address at least the more important points in [GRChombo/GRChombo#187](https://github.com/GRChombo/GRChombo/issues/187), I will recommend this work to publication in JOSS.

Yes, I completely agree with these comments. We are going to update the main points regarding the wiki and will ask you to take another look once this is done. We will probably leave an issue outstanding for the more minor points remaining and tackle them more slowly as and when we have time, but at least we will have them listed somewhere for reference.

Thanks again for your really helpful feedback and for your hard work and persistence in getting this to work :-)
