### 2021-11-19 5ca50df
Add WIP CatalystAdaptor class
