### 2021-11-22 GRChombo/GRChombo#187 @KAClough
I think these have all now been addressed or moved to specific issues so I am closing this issue.
