+++ README.md
@@ -1,5 +1,5 @@
 # GRChombo
-GRChombo is a new open-source code for numerical general relativity simulations.
+GRChombo is an open-source code for numerical relativity simulations.
 It is developed and maintained by a collaboration of numerical relativists with a
 wide range of research interests, from early universe cosmology to astrophysics
 and mathematical general relativity, and has been used in many papers since its
@@ -16,12 +16,7 @@ institutions, a list of publications using GRChombo, and some videos.
 
 ## Getting started
 Detailed installation instructions and usage examples are available in
-our [wiki](https://github.com/GRChombo/GRChombo/wiki).
-
-## IMPORTANT: Updating examples for diagnostic variables
-If you wish to merge diagnostic variables changes in this branch into an older
-branch, you will need to update old examples. Please consult
-[this wiki page](https://github.com/GRChombo/GRChombo/wiki/Updating-old-examples-for-Diagnostic-Variables-changes).
+our [wiki](https://github.com/GRChombo/GRChombo/wiki), with the home page giving guidance on where to start.
 
 ## Contributing
 We welcome feedback, bug reports, and contributions. Please consult the [wiki](https://github.com/GRChombo/GRChombo/wiki)
