### 2021-11-22 #3703 @Sbozzolo
> @Sbozzolo sorry for the delay, I have now updated the major points you made regarding the wiki, in particular:
>
> * The home page https://github.com/GRChombo/GRChombo/wiki now gives a lot more guidance on how to use the wiki, and how to report bugs, contribute (I have simply linked to this in the README.md file, rather than duplicate wording from the wiki).
> * There is a new Capabilities page: https://github.com/GRChombo/GRChombo/wiki/Capabilities which details current and planned features.
> * The structure of the code page is up to date: https://github.com/GRChombo/GRChombo/wiki/Structure-of-the-code-%28aka-where-is-everything%3F%29 as are the details that follow on how the Scalar example was written.
> * Some guidance on convergence testing is given https://github.com/GRChombo/GRChombo/wiki/Convergence-testing
> * @mirenradia has fixed the oxygen docs and added instructions on how to build them here: https://github.com/GRChombo/GRChombo/wiki/Doxygen-documentation
>
> There are still a few pages that need more minor updates - we have made separate issues for them so they get addressed in the longer term, but I think the wiki is already hugely improved for these amendments. Thanks again for the suggestions!
>
> Please take a look and let us know if there are any key points outstanding.

This is terrific work! I think that the documentation improved dramatically. My final (non-blocking) comments are:
- In the documentation, you ask users to report bug `detailing the exact version of the code you are using`. If this identified with the commit hash?
- I would think about the copyright more carefully, even it is something that is not typically important in software for numerical relativity. Who is "GRChombo collaboration"? If I contribute to GRChombo, do I own my contribution?
- The about of the GitHub repo says "Main public version of the GRChombo code". To me, the implied message is that there's an another repo where the real development happens. You can change with something more descriptive like: "A flexible, modern, numerical-relativity code".
- I highly encourage you to think about your post-processing tools next.

With this, I am going to recommend this paper to publication (@rkurchin).

Thanks, @KAClough, @mirenradia and the GRChombo collaborators for releasing the code. I hope you all keep up with this great work and maintain a high level of quality, openness, and documentation.
