### 2021-11-23 #3703 @rkurchin
Fantastic! Thanks everyone here for your extensive efforts on a review that has clearly made this software better and easier to use!

@KAClough, I'll let you decide if you want to make any changes corresponding to the last few comments above. Next steps will be for me to make an editorial pass over the manuscript to be another set of eyes for typographical/style errors, etc. and then we should be ready to rock! Given the holiday this week, I will plan on doing this early next week.

Congratulations!
