### 2021-11-23 #3703 @KAClough
Thanks @rkurchin, @Sbozzolo and @rashti-alireza, we did really find this a useful review process :-)

I have updated for the 1st and 3rd bullets (the 2nd and 4th will take a bit more time and discussion amongst the collaboration to resolve), so I think everything is good for you @rkurchin to go over the draft next week or whenever you have time. Happy Thanksgiving!
