### 2021-11-23 8bf6560
Add Make.insitu
This file allows GRChombo examples to be built with Catalyst. Also add
some more work on the CatalystAdaptor code.
