+++ Source/Insitu/CatalystAdaptor.hpp
@@ -11,6 +11,9 @@
 // Standard library includes
 #include <string>
 
+// Chombo includes
+#include "MayDay.H"
+
 // GRChombo includes
 #include "GRAMR.hpp"
 #include "UserVariables.hpp"
@@ -26,6 +29,12 @@
 #include <vtkNew.h>
 #include <vtkOverlappingAMR.h>
 #include <vtkUniformGrid.h>
+
+// Chombo namespace
+#include "UsingNamespace.H"
+
+// Forward declaration of GRAMR class
+class GRAMR;
 
 // Class that interfaces with ParaView Catalyst for insitu visualisation
 class CatalystAdaptor
@@ -57,7 +66,6 @@ class CatalystAdaptor
 
   private:
     bool m_initialised = false;
-    bool m_finalised = false;
     const GRAMR *m_gr_amr_ptr = nullptr;
     std::string m_python_script_path;
 
