### 2021-11-30 276c3d9
Add script to cache output from paraview-config
paraview-config is extremely slow and we only want to call it once and
use the cached output otherwise.
