### 2021-11-30 1ab1957
Add CatalystAdaptor to GRAMR
Also add compile_commands.json database.
