+++ Source/GRChomboCore/GRAMR.cpp
@@ -8,6 +8,14 @@
 
 GRAMR::GRAMR() : m_interpolator(nullptr) {}
 
+void GRAMR::conclude()
+{
+    AMR::conclude();
+#ifdef USE_CATALYST
+    m_insitu->finalise();
+#endif
+}
+
 // Called after AMR object set up
 void GRAMR::set_interpolator(AMRInterpolator<Lagrange<4>> *a_interpolator)
 {
@@ -51,3 +59,12 @@ void GRAMR::fill_multilevel_ghosts(const VariableType a_var_type,
         level.fillAllGhosts(a_var_type, a_comps);
     }
 }
+
+#ifdef USE_CATALYST
+void GRAMR::setup_catalyst(const std::string &a_python_script_path)
+{
+    pout() << "GRAMR::setup_catalyst" << std::endl;
+    m_insitu = new CatalystAdaptor;
+    m_insitu->initialise(this, a_python_script_path);
+}
+#endif
