### 2021-12-01 cd576ce
Add CatalystAdaptor::build_vtk_grid function
This currently passes ghosted boxes to Catalyst.
