### 2021-12-01 8707946
Add CatalystAdaptor::add_vars
In theory only the variables requested in the script should be sent to
Catalyst.
