### 2021-12-07 ae6cddc
Implement CatalystAdaptor::coprocess and call it
