+++ paper.md
@@ -144,21 +144,21 @@ Strong gravity regimes are described by the *Einstein Field Equation* (EFE) of G
    R_{\mu \nu} - \frac{1}{2} R g_{\mu \nu} = 8 \pi G T_{\mu\nu} ~,
 \end{equation}
 where $g_{\mu\nu}$ is the gravitational metric describing spacetime distances, $R$ and $R_{\mu\nu}$ are related to its second derivatives in space and time, and $T_{\mu\nu}$ is the stress-energy tensor of any matter or fields present.
-Analytic solutions to the EFE only exist where there is a high degree of symmetry; in general the equations must be solved numerically. The need for observational predictions has thus led to the development of *numerical relativity* (NR), methods for numerically solving the above expression, typically utilising high performance computing (HPC) resources.
-Expanding out the tensorial notation above, the EFE is a set of coupled, non linear second order partial differential equations for $g_{\mu\nu}$, which describes the curvature of spacetime in the presence of stress-energy $T_{\mu\nu}$, that is, schematically the equation we are trying to solve has the form:
+Analytic solutions to the EFE only exist where there is a high degree of symmetry; in general the equations must be solved numerically. The need for observational predictions has thus led to the development of *numerical relativity* (NR), methods for numerically solving the above expression, typically utilising high-performance computing (HPC) resources.
+Expanding out the tensorial notation above, the EFE is a set of coupled, nonlinear second-order partial differential equations for $g_{\mu\nu}$, which describes the curvature of spacetime in the presence of stress-energy $T_{\mu\nu}$, that is, schematically the equation we are trying to solve has the form:
 \begin{equation}
-   \partial_t \partial_t g_{\mu\nu} \sim \partial_x \partial_x g_{\mu\nu} + \partial_y \partial_y g_{\mu\nu} + \partial_z \partial_z g_{\mu\nu} + {\rm non ~ linear ~ cross ~ terms} 
+   \partial_t \partial_t g_{\mu\nu} \sim \partial_x \partial_x g_{\mu\nu} + \partial_y \partial_y g_{\mu\nu} + \partial_z \partial_z g_{\mu\nu} + {\rm nonlinear ~ cross ~ terms} 
           + 8 \pi G T_{\mu\nu} ~,
 \end{equation}
-where the indices $\mu, \nu$ run over the spacetime indices -- in 4 dimensions, $t, x, y, z$. Given that $g_{\mu\nu}$ is symmetric in its indices, this gives a set of ten coupled non-linear partial differential equations, sourced by the stress-energy of any matter or fields present in the spacetime.
+where the indices $\mu, \nu$ run over the spacetime indices -- in 4 dimensions, $t, x, y, z$. Given that $g_{\mu\nu}$ is symmetric in its indices, this gives a set of ten coupled nonlinear partial differential equations, sourced by the stress-energy of any matter or fields present in the spacetime.
 
[…]
