### 2021-12-08 #3703 @KAClough
Hi @rkurchin sorry for being slow - I've made the updates to the file as requested but I am just looking at the Zenodo/figshare thing and canvassing opinions on which we prefer as a team. I guess I should move the Joss paper into the main branch before I tag the release, right?
