### 2021-12-08 #3703 @rkurchin
No worries!

You should definitely make sure any associated changes from the discussions here are merged. If you want the paper to live in the repo, you can definitely merge that too but it’s not required (you can always just put a badge to link to the paper that we host if you don’t feel like having the manuscript always be in the repo).
