### 2021-12-10 #3703 @rkurchin
Oops, didn't realize I needed to do it for the accept too, but that makes perfect sense, my bad!

@KAClough can you fix that one missing DOI?
