+++ paper.bib
@@ -603,7 +603,11 @@ booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific 
     eprint = "2006.09414",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
-    month = "6",
+    doi = "10.1088/1361-6382/abb693",
+    journal = "Class. Quant. Grav.",
+    volume = "37",
+    number = "22",
+    pages = "225009",
     year = "2020"
 }
 
