### 2021-12-10 #3703 @danielskatz
Hi @KAClough - the way the captions and the images fit together in the pdf isn't quite right. I wonder if you could try adding a blank link between the image and the caption, for example in line 195.  Also, some of the cases in the bib entries aren't right, for example, "Universe" in Adamek should be "universe", and Andrade should be "Violations of weak cosmic ..."

Can you make appropriate changes for this?
