### 2021-12-10 1e97143
Update bib and image spaces
