+++ paper.bib
@@ -145,7 +145,7 @@
 
 @article{Arnowitt:1962hi,
     author = "Arnowitt, Richard L. and Deser, Stanley and Misner, Charles W.",
-    title = "{The Dynamics of general relativity}",
+    title = "{The dynamics of general relativity}",
     eprint = "gr-qc/0405109",
     archivePrefix = "arXiv",
     doi = "10.1007/s10714-008-0661-1",
@@ -249,7 +249,7 @@
 
 @article{Clough:2015sqa,
     author = "Clough, Katy and Figueras, Pau and Finkel, Hal and Kunesch, Markus and Lim, Eugene A. and Tunyasuvunakool, Saran",
-    title = "{GRChombo : Numerical Relativity with Adaptive Mesh Refinement}",
+    title = "{GRChombo : Numerical relativity with adaptive mesh refinement}",
     eprint = "1503.03436",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
@@ -275,7 +275,7 @@
 
 @article{Pfeiffer:2002wt,
     author = "Pfeiffer, Harald P. and Kidder, Lawrence E. and Scheel, Mark A. and Teukolsky, Saul A.",
-    title = "{A Multidomain spectral method for solving elliptic equations}",
+    title = "{A multidomain spectral method for solving elliptic equations}",
     eprint = "gr-qc/0202096",
     archivePrefix = "arXiv",
     doi = "10.1016/S0010-4655(02)00847-0",
@@ -367,7 +367,7 @@ booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific 
 
 @article{Adamek:2020jmr,
     author = "Adamek, Julian and Barrera-Hinojosa, Cristian and Bruni, Marco and Li, Baojiu and Macpherson, Hayley J. and Mertens, James B.",
-    title = "{Numerical solutions to Einstein's equations in a shearing-dust Universe: a code comparison}",
+    title = "{Numerical solutions to Einstein's equations in a shearing-dust universe: a code comparison}",
     eprint = "2003.08014",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
@@ -470,7 +470,7 @@ booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific 
 
 @article{Brugmann:2008zz,
     author = "Bruegmann, Bernd and Gonzalez, Jose A. and Hannam, Mark and Husa, Sascha and Sperhake, Ulrich and Tichy, Wolfgang",
-    title = "{Calibration of Moving Puncture Simulations}",
+    title = "{Calibration of moving puncture simulations}",
     eprint = "gr-qc/0610128",
     archivePrefix = "arXiv",
     doi = "10.1103/PhysRevD.77.024027",
@@ -482,7 +482,7 @@ booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific 
 
 @article{Ruchlin:2017com,
     author = "Ruchlin, Ian and Etienne, Zachariah B. and Baumgarte, Thomas W.",
-    title = "{SENR/NRPy+: Numerical Relativity in Singular Curvilinear Coordinate Systems}",
+    title = "{SENR/NRPy+: Numerical relativity in singular curvilinear coordinate systems}",
     eprint = "1712.07658",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
@@ -599,7 +599,7 @@ booktitle = {{High Performance Visualization--Enabling Extreme-Scale Scientific 
 
 @article{Figueras:2020dzx,
     author = "Figueras, Pau and Fran\c{c}a, Tiago",
-    title = "{Gravitational Collapse in Cubic Horndeski Theories}",
[…]
