### 2021-12-10 #3703 @danielskatz
Thanks - for what it's worth in the bib entries, you are using `"{entry}"` rather than just `{entry}`, which makes your entries be used literally, rather than bibtex doing the work.  As long as they right, it doesn't really matter, but in general, it would be better to let the style be figured out by the tools.
