### 2021-12-13 #3703 @KAClough
@danielskatz and @rkurchin we have just realised that a couple of acknowledgements of funding were missing from the paper - they had been added earlier but got lost in a rebase. Is there any possibility of updating the published copy for these? I have added them to the branch in this commit:
https://github.com/GRChombo/GRChombo/compare/1e97143a644b...ad4d7aec747f

I'm very sorry for this, I should have noticed it earlier. We'd really appreciate if you could suggest a way to include them, since the funders are quite fussy about such things and this is an important publication for us.
