### 2022-01-10 #3703 @KAClough
> 👋 @openjournals/dev - can you handle this?

Sorry to chase @openjournals/dev, but was this ok? There is no rush, but it would be great if it can be updated. Thanks!
