### 2022-01-10 #3703 @arfon
@KAClough – I've updated the paper with your changes. The PDF might take ~24 hours to show as updated on the JOSS site due to caching.
